requirements.txt
# requirements.txt
PyQt5>=5.15.0
pyusb>=1.2.0
watchdog>=2.1.6
yara-python>=4.2.0
psutil>=5.9.0